requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.3
cssselect==1.1.0
cssutils==2.6.0
distro==1.8.0
dj-rest-auth==4.0.1
Django==4.2.4
django-appconf==1.0.5
django-cors-headers==4.2.0
django-cryptography==1.1
django-extensions==3.2.3
django-trench==0.3.1
djangorestframework==3.14.0
Faker==19.3.1
frozenlist==1.3.1
geoip2==4.7.0
gunicorn==21.2.0
idna==3.4
invoke==2.2.0
lxml==4.9.2
maxminddb==2.4.0
multidict==6.0.2
packaging==23.1
premailer==3.10.0
psutil==5.9.5
psycopg==3.1.10
pycparser==2.21
PyJWT==2.5.0
pymailq==0.9.0
pyotp==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.31.0
six==1.16.0
smsapi-client==2.6.0
sqlparse==0.4.4
twilio==7.14.1
typing_extensions==4.3.0
urllib3==2.0.2
yarl==1.8.1
yubico-client==1.13.0
